fix : 추가하기 모달창 수정
